+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All Notable changes to `League\Uri\Components` will be documented in this file
 
+## Next
+
+### Added
+
+- None
+
+### Fixed
+
+- None
+
+### Deprecated
+
+- None
+
+### Removed
+
+- PHP5 support
+
 ## 0.4.0 - 2016-12-09
 
 ### Added