+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All Notable changes to `League\Uri\Components` will be documented in this file
 
+## 1.0.3 - TBD
+
+### Added
+
+- None
+
+### Fixed
+
+- Bug fix `$_SERVER['SERVER_PORT']` valui with `League\Uri\Schemes\Http::createFromServer` see [#PR1](https://github.com/thephpleague/uri-schemes/pull/1)
+
+### Deprecated
+
+- None
+
+### Removed
+
+- None
+
 ## 1.0.2 - 2017-03-01
 
 ### Added